% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{meridian_version}
\alias{meridian_version}
\title{Installed Meridian version}
\usage{
meridian_version()
}
\value{
A string with the Meridian Python package version.
}
\description{
Installed Meridian version
}
