% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meridianR-package.R
\docType{package}
\name{meridianR-package}
\alias{meridianR}
\alias{meridianR-package}
\title{meridianR: R Interface to Google Meridian (Marketing Mix Modeling)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R interface to Google's 'Meridian' Bayesian Marketing Mix Modeling (MMM) library. Wraps the 'Meridian' 'Python' package via 'reticulate', providing faithful bindings to its data, model, and analysis objects together with an idiomatic R layer for loading data, configuring priors and model specifications, fitting models, and analyzing results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Roeh-Marketing/meridianR}
  \item \url{https://roeh-marketing.github.io/meridianR/}
  \item Report bugs at \url{https://github.com/Roeh-Marketing/meridianR/issues}
}

}
\author{
\strong{Maintainer}: Ben Denis Shaffer \email{shafferbendenis@gmail.com}

}
\keyword{internal}
