% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hill_curves}
\alias{plot_hill_curves}
\title{Plot Hill saturation curves}
\usage{
plot_hill_curves(model, ...)
}
\arguments{
\item{model}{A fitted \code{meridian_model} (or an \code{\link[=analyzer]{analyzer()}}).}

\item{...}{Passed to \code{\link[=hill_curves]{hill_curves()}}.}
}
\value{
A ggplot object.
}
\description{
Plot Hill saturation curves
}
\seealso{
Other meridian plots: 
\code{\link{plot_adstock_decay}()},
\code{\link{plot_contribution}()},
\code{\link{plot_model_fit}()},
\code{\link{plot_response_curves}()},
\code{\link{plot_rhat}()},
\code{\link{plot_roi}()}
}
\concept{meridian plots}
