% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_roi}
\alias{plot_roi}
\title{Plot return on investment by channel}
\usage{
plot_roi(model, confidence_level = 0.9, ...)
}
\arguments{
\item{model}{A fitted \code{meridian_model} (or an \code{\link[=analyzer]{analyzer()}}).}

\item{confidence_level}{Credible-interval width for the error bars.}

\item{...}{Passed to \code{\link[=summary_metrics]{summary_metrics()}}.}
}
\value{
A ggplot object.
}
\description{
Plot return on investment by channel
}
\seealso{
Other meridian plots: 
\code{\link{plot_adstock_decay}()},
\code{\link{plot_contribution}()},
\code{\link{plot_hill_curves}()},
\code{\link{plot_model_fit}()},
\code{\link{plot_response_curves}()},
\code{\link{plot_rhat}()}
}
\concept{meridian plots}
