% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{import}
\alias{py_require}
\alias{py_config}
\alias{r_to_py}
\alias{py_to_r}
\alias{use_python}
\alias{use_virtualenv}
\alias{use_condaenv}
\alias{fit}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{reticulate}{\code{\link[reticulate]{import}}, \code{\link[reticulate]{py_config}}, \code{\link[reticulate]{py_require}}, \code{\link[reticulate:r-py-conversion]{py_to_r}}, \code{\link[reticulate:r-py-conversion]{r_to_py}}, \code{\link[reticulate:use_python]{use_condaenv}}, \code{\link[reticulate]{use_python}}, \code{\link[reticulate:use_python]{use_virtualenv}}}
}}

